tests/integration/requirements.txt
pytest==8.3.4
psycopg2-binary==2.9.10
requests==2.32.3
